% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_draw_MVT.R
\name{compute_MC_draws_mvt}
\alias{compute_MC_draws_mvt}
\title{Compute Monte Carlo Draws from Multivariate Dataset}
\usage{
compute_MC_draws_mvt(
  Data,
  tau,
  At_lens,
  B,
  nu0,
  V0 = mapply(diag, p_list, SIMPLIFY = FALSE),
  alph,
  gam,
  p_list
)
}
\arguments{
\item{Data}{Observed data organised as a list of {y, X, A} where y is a vector of the final outcomes,
X is a list of matrices of the intermediate covariates
and A is a matrix of the assigned treatments}

\item{tau}{Prior precision scale. Should be specified with a small value}

\item{At_lens}{Vector of number of treatment options at each stage}

\item{B}{Number of MC draws}

\item{nu0}{Inverse-Wishart degres of freedom}

\item{V0}{Inverse-Wishart scale matrix}

\item{p_list}{Vector of dimension for each stage}
}
\value{
Monte Carlo draws??? A list containing:  \enumerate{
\item sigmat_B_list: Desc. A list of length T with each element a vector of size B x p_t
\item Wt_B_list: Desc. A list of length T with each element a matrix of size B x p_t
\item beta_B: Desc. A list of length B
\item sigmay_2B: Desc. A list of length B
}
}
\description{
Obtain Monte Carlo draws from posterior distribution of stagewise regression parameters
}
\examples{
# -----------------------------
# Initialise Inputs
# -----------------------------
n           <- 5000
num_treats  <- 3
p_list      <- rep(2, num_treats)
At_lens     <- rep(3, num_treats)
Data        <- generate_dataset_mvt(n, num_treats, p_list, At_lens)

# -----------------------------
# Main
# -----------------------------
res_mvt <- compute_MC_draws_mvt(Data = Data, tau = 0.01, At_lens = 3, B = 100, nu0 = 3, V0 = diag(2), alph = 3, gam = 4, p_list = 2)
}
