% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{generate_dataset}
\alias{generate_dataset}
\title{Generate Dataset}
\usage{
generate_dataset(n, num_stages, p_list, num_treats)
}
\arguments{
\item{n}{Number of samples/individuals to generate}

\item{num_stages}{Total number of stages per individual}

\item{p_list}{Vector of dimension for each stage}

\item{num_treats}{Vector of number of treatment options at each stage}
}
\value{
Observed data organised as a list of \eqn{\{y, X_1, X_2..., X_{num\_stages}, A\}} where y is a
vector of the final outcomes, \eqn{X_1, X_2..., X_{num\_stages}} is a list of matrices
of the intermediate covariates and A is an \eqn{n \times num\_stages}{n x num_stages} matrix of the
assigned treatments
}
\description{
Generate a toy dataset for testing BayesLinRegDTR.model.fit
}
\examples{
# -----------------------------
# Initialise Inputs
# -----------------------------
n           <- 5000
num_stages  <- 3
p_list      <- rep(2, num_stages)
num_treats  <- rep(3, num_stages)

# -----------------------------
# Main
# -----------------------------
Data        <- generate_dataset(n, num_stages, p_list, num_treats)
}
