% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen_univariate.R
\name{generate_dataset_uvt}
\alias{generate_dataset_uvt}
\title{Generate Univariate Dataset}
\usage{
generate_dataset_uvt(n = 5000, num_treats = 5, At_len = 3)
}
\arguments{
\item{n}{Number of samples to generate}

\item{num_treats}{Total number of stages per sample}

\item{At_len}{Vector of number of treatment options at each stage}
}
\value{
Observed data organised as a list of {y, X, A} where y is a vector of the final outcomes,
X is a list of matrices of the intermediate covariates
and A is a matrix of the assigned treatments
}
\description{
Generate Univariate Dataset
}
\examples{
# -----------------------------
# Initialise Inputs
# -----------------------------
n           <- 500
num_treats  <- 5
At_len      <- 3

# -----------------------------
# Main
# -----------------------------
Data <- generate_dataset_uvt(n, num_treats, At_len)

}
