% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestParallelGCV.R
\name{testParallelGCV}
\alias{testParallelGCV}
\title{Title}
\usage{
testParallelGCV(
  Dat.train,
  Dat.pred,
  n.train,
  n.pred,
  num_stages,
  num_treats,
  p_list,
  t,
  R = 30,
  numCores,
  tau = 0.01,
  B = 10000,
  nu0 = 3,
  V0 = mapply(diag, p_list, SIMPLIFY = FALSE),
  alph = 1,
  gam = 1
)
}
\arguments{
\item{Dat.train}{Training data in format returned by \code{generate_dataset}:
organised as a list of \eqn{\{y, X_1, X_2..., X_{num\_states}, A\}} where y is a
vector of the final outcomes, \eqn{X_1, X_2..., X_{num\_states}} is a list of matrices
of the intermediate covariates and A is an \eqn{n \times num\_stages}{n x num_stages} matrix of the
assigned treatments}

\item{Dat.pred}{Prediction data in format returned by \code{generate_dataset}:
organised as a list of \eqn{\{X_1, X_2..., X_t, A\}} where
\eqn{X_1, X_2..., X_t} is a list of matrices of the intermediate
covariates and A is an \eqn{n \times (t-1)}{n x (t-1)} matrix of the assigned treatments}

\item{n.train}{Number of samples/individuals in the training data}

\item{n.pred}{Number of samples/individuals in the prediction data}

\item{num_stages}{Total number of stages}

\item{num_treats}{Vector of number of treatment options at each stage}

\item{p_list}{Vector of intermediate covariate dimensions for each stage}

\item{t}{Prediction stage t, where \eqn{\leq}{<=} num_stages}

\item{R}{Draw size from posterior distribution of intermediate covariates. default:  30}

\item{numCores}{Number of cores in the system to use. default uses parallel::detectCores() - 1}

\item{tau}{Normal prior scale parameter for regression coefficients. Should be specified with a small value. default:  0.01}

\item{B}{Number of MC draws from posterior of regression parameters. default 10000}

\item{nu0}{Inverse-Wishart prior degrees of freedom for regression error Vcov matrix. default: 3}

\item{V0}{List of Inverse-Wishart prior scale matrix for regression error Vcov matrix. default: list of identity matrices}

\item{alph}{Inverse-Gamma prior shape parameter for regression error variance of y. default:  1}

\item{gam}{Inverse-Gamma prior rate parameter for regression error variance of y. default:  1}
}
\value{
list of {GCV_results, freqs} where GCV_results is an array of size
n.pred x p_list[t] x B, and freqs is a n.pred x p_list[t] matrix
of the frequency of each treatment type at stage t
}
\description{
Title
}
\examples{
# -----------------------------
# Initialise Inputs
# -----------------------------
num_stages  <- 3
p_list      <- rep(2, num_stages)
num_treats  <- rep(2, num_stages)
n.train     <- 5000
n.pred      <- 100

tau     <- 0.01
B       <- 100
alph    <- 3
gam     <- 4
nu0     <- 3
V0      <- mapply(diag, p_list, SIMPLIFY = FALSE)
R       <- 30
numCores<- parallel::detectCores()

# -----------------------------
# Generate Dataset
# -----------------------------
Dat.train <- generate_dataset_mvt(n.train, num_stages, p_list, num_treats)
Dat.pred  <- generate_dataset_mvt(n.pred,  num_stages, p_list, num_treats)
Dat.pred  <- Dat.pred[-1]
Dat.pred[[num_stages+1]]  <- Dat.pred[[num_stages+1]][1:n.pred, 1:(t-1), drop = FALSE]

# -----------------------------
# Main
# -----------------------------
gcv_res <- testParallelGCV(Dat.train, Dat.pred, n.train, n.pred, num_stages, num_treats, p_list, t)
}
