% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_draw_large.R
\name{compute_MC_draws_mvt}
\alias{compute_MC_draws_mvt}
\title{Title}
\usage{
compute_MC_draws_mvt(
  D,
  tau,
  At_lens,
  B,
  nu0,
  V0 = mapply(diag, p_list, SIMPLIFY = FALSE),
  p_list
)
}
\arguments{
\item{p_list}{}
}
\description{
Title
}
