% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen_multivariate.R
\name{generate_dataset_mvt}
\alias{generate_dataset_mvt}
\title{Generate Multivariate dataset}
\usage{
generate_dataset_mvt(n, num_treats, p_list, At_len)
}
\arguments{
\item{n}{Number of samples to generate}

\item{num_treats}{Total number of stages per sample}

\item{p_list}{Vector of dimension for each stage}

\item{At_len}{Vector of number of treatment options at each stage}
}
\value{
Observed data organised as a list of {y, X, A} where y is a vector of the final outcomes,
X is a list of matrices of the intermediate covariates
and A is a matrix of the assigned treatments
}
\description{
Generate Multivariate dataset
}
\examples{
# -----------------------------
# Initialise Inputs
# -----------------------------
n           <- 5000
num_treats  <- 3
p_list      <- rep(2, num_treats)
At_len      <- 3

# -----------------------------
# Main
# -----------------------------
Data        <- generate_dataset_mvt(n, num_treats, p_list, At_len)
}
