% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen_multivariate.R
\name{generate_dataset_mvt}
\alias{generate_dataset_mvt}
\title{Generate Multivariate dataset}
\usage{
generate_dataset_mvt(n, num_stages, p_list, num_treats)
}
\arguments{
\item{n}{Number of samples/individuals to generate}

\item{num_stages}{Total number of stages per individual}

\item{p_list}{Vector of dimension for each stage}

\item{num_treats}{Vector of number of treatment options at each stage}
}
\value{
Observed data organised as a list of \eqn{\{y, X, A\}} where y is a
vector of the final outcomes, X is a list of matrices of the intermediate
covariates and A is a matrix of the assigned treatments
}
\description{
Generate Multivariate dataset
}
\keyword{internal}
