% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen_univariate.R
\name{generate_dataset_uvt}
\alias{generate_dataset_uvt}
\title{Generate Univariate Dataset}
\usage{
generate_dataset_uvt(n, num_stages, num_treats)
}
\arguments{
\item{n}{Number of samples/individuals to generate}

\item{num_stages}{Total number of stages per individual}

\item{num_treats}{Vector of number of treatment options at each stage}
}
\value{
Observed data organised as a list of \eqn{\{y, X, A\}} where y is a
vector of the final outcomes, X is a list of matrices of the intermediate
covariates and A is a matrix of the assigned treatments
}
\description{
Generate Univariate Dataset
}
\keyword{internal}
